+++ lib/redox/models/patient.rb
@@ -3,7 +3,11 @@
 require_relative "patient_identifier"
 require_relative "demographics"
 
-class Redox::Models::Patient < Redox::Model
-  redox_property :Identifiers, coerce: Array[Redox::Models::PatientIdentifier]
-  redox_property :Demographics, coerce: Redox::Models::Demographics
+module Redox
+  module Models
+    class Patient < Redox::Model
+      redox_property :Identifiers, coerce: Array[PatientIdentifier]
+      redox_property :Demographics, coerce: Demographics
+    end
+  end
 end