+++ lib/redox/models/patient_identifier.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 
-class Redox::Models::PatientIdentifier < Redox::Model
-  redox_property :IDType
-  redox_property :ID
+module Redox
+  module Models
+    class PatientIdentifier < Redox::Model
+      redox_property :IDType
+      redox_property :ID
+    end
+  end
 end